==> Xray/Pre-process/0_parse_data_folder.ipynb <==
import os
import glob
from pathlib import Path
import json
import pandas as pd
import numpy as np
import nibabel as nib
from PIL import Image
import matplotlib.pyplot as plt

wd = Path.cwd().parent
data_dir = '/project/dane2/wficai/pda'
os.listdir(data_dir)

os.listdir(data_dir)

metadata_glob = data_dir + '/Images/*/*.json'
nifti_glob = data_dir + '/Images/*/*.nii'

csv_metadata_path = (wd / 'data/metadata.csv').resolve().as_posix()
csv_metadata_path

csv_nifti_path = (wd / 'data/nifti.csv').resolve().as_posix()

png_export_path = (wd / 'data/exported_pngs').resolve().as_posix()
jpg_export_path = (wd / 'data/exported_jpgs').resolve().as_posix()
png_export_path, jpg_export_path

## Utilities

def clean_pid(pid):
    an_num = pid.split(' ')[-1].replace('(', '').replace(')', '')
    pid_clean = 'AN' + f"{an_num.zfill(5)}"
    
    return pid_clean
    
def next_power_of_2(x):  
    return 1 if x == 0 else 2**(x - 1).bit_length()

## Load metadata

meta_data_files = glob.glob(metadata_glob, recursive=True)

meta_data_files


len(meta_data_files)

def read_json(filepath):
    with open(filepath, 'r') as fp:
        record = json.load(fp)
    
    record['FULL_PATH'] = filepath.replace('\\', '/')
        
    # add metadata embedded in filepath
    path_split = filepath.split('/')
    record['PATIENT_ID'] = clean_pid(path_split[-2])
    record['FILENAME_VIEW'] = path_split[-1].split('.json')[0]
    
    return record

read_json(meta_data_files[2])

meta_data = pd.DataFrame([read_json(fp) for fp in meta_data_files])

meta_data

meta_data.to_csv(csv_metadata_path, index=None)

## Load images

nifti_files = [fp.replace('\\', '/') for fp in glob.glob(nifti_glob, recursive=True)]

len(nifti_files)

def read_nifti(filepath):
    path_split = filepath.split('/')
    record = dict()
    record['PATIENT_ID'] = clean_pid(path_split[-2])
    record['FULL_PATH'] = filepath
    record['FILENAME_VIEW'] = path_split[-1].split('.nii')[0]
    
    #load the nifti image
    img_nib = nib.load(filepath)
    img = np.asarray(img_nib.dataobj)
    record['MIN'] = img.min()
    record['MAX'] = img.max()
    record['DATA_TYPE'] = img.dtype
    record['WIDTH'] = img.shape[-3]
    record['HEIGHT'] = img.shape[-2]
    record['NUM_CHANNELS'] = img.shape[-1]
    
    # determine if image represents a ROI
    record['ROI'] = 'ROI' in record['FILENAME_VIEW'].split('_')[-1]
    
    # Format image for saving
    img = img.astype(np.float32)
    img = (img - record['MIN']) / (record['MAX'] - record['MIN'])
    img = img.squeeze().T[-1::-1]
    
    # save png and jpg
    save_path_jpg =  jpg_export_path + f"/{record['PATIENT_ID']}_{record['FILENAME_VIEW']}.jpg"
    record['JPG_PATH'] = save_path_jpg
    plt.imsave(save_path_jpg, img, cmap='Greys_r')
    
    return record

nifti_files[0]

nifti_meta_data = []
for ix, fp in enumerate(nifti_files):
    try:
        print(f'Processing image {ix+1} / {len(nifti_files)}', end='\r')
        nifti_meta_data.append(read_nifti(fp))
    except Exception as e:
        print(e)

nifti_meta_data = pd.DataFrame(nifti_meta_data)   

nifti_meta_data

## save

nifti_meta_data.to_csv(csv_nifti_path, index=None)

csv_nifti_path
