==> Xray/Pre-process/1_make_lbx_data_json.ipynb <==
from pathlib import Path
import json
import pandas as pd
from PIL import Image

wd = Path.cwd().parent
data_dir = (wd / 'data/Rad_images').resolve()

data_dir

csv_nifti_path = (wd / 'data/nifti.csv').resolve().as_posix()
jpg_export_path = (wd / 'data/exported_jpgs').resolve().as_posix()
lbx_path = (wd / 'data/lbx_data.json').resolve().as_posix()

df = pd.read_csv(csv_nifti_path)

df

# Create data json for labelbox

def create_record(row):
    filename = row['JPG_PATH'].split('/')[-1]

    record = dict(
        externalId = filename[:-4],
        imageUrl = f"https://fast-videos.s3.us-east-2.amazonaws.com/phf/cobra_prelim_set/{filename}"
    )
    
    return record
    
create_record(next(df.iterrows())[1])

print(df.columns)

df['ROI']

records = []
for ix, row in df.iterrows():
    # if not row['ROI'] and row['is_valid']:
    if not row['ROI']: 
        records.append(create_record(row))

records[:5]

len(records)

with open(lbx_path, 'w') as f:
    json.dump(records, f)
