==> Xray/Pre-process/2_crop_to_roi.ipynb <==
import json
import pandas as pd
from PIL import Image
import numpy as np

roi_data_path = '../data/roi_export-2022-09-20T01 40 52.606Z.json' 
nifti_meta_path = '../data/nifti.csv'
cropped_image_path = '../data/cropped_images/'
cropped_meta_path = '../data/cropped_image_metadata.csv'

# Parse label json objects

with open(roi_data_path, 'r') as f:
    data = json.load(f)

roi_data_path


data[0]

data[0]['Label']

def parse_roi(lab):
    
    result = dict()
    
    # get external id
    result['external_id'] = lab['External ID']
    
    # get bounding box
    objects = lab['Label']['objects']
    if len(objects)>1:
        print("Warning: more than one bounding box. Using the first box.")    
        bbox = objects[0]['bbox']
    elif len(objects)==0:
        bbox=None
    else:
        bbox = objects[0]['bbox']
    
    if bbox is not None:
        for k,v in bbox.items():
            result[f'bbox_{k}'] = v
    
    # get classifications:
    for c in lab['Label']['classifications']:
        if c['value'] == 'special_notes':
            for a in c['answers']:
                result[a['value']] = True
            
        elif c['value'] == 'optional_notes':
            result['note'] = c['answer']
        
    return result

df = pd.DataFrame([parse_roi(d) for d in data])
df

# impute False flags
df.brightness_flipped = df.brightness_flipped.fillna(False)
df.poor_quality = df.poor_quality.fillna(False)
df.not_a_phf_view = df.not_a_phf_view.fillna(False)

# drop rows with missing bbox
df = df.dropna(subset=['bbox_top', 'bbox_left', 'bbox_height', 'bbox_width'])

df[~df.note.isna()]

df.brightness_flipped.value_counts()

df.poor_quality.value_counts()

df.not_a_phf_view.value_counts()

# join on image data

df_nifti = pd.read_csv(nifti_meta_path)
df_nifti['external_id'] = df_nifti['PATIENT_ID'] + '_' + df_nifti['FILENAME_VIEW']

df_wide = df.merge(df_nifti, how='inner', on='external_id')
assert len(df_wide) == len(df), "Data loss when merging nifti metadata with ROI data. ROI data should be a subset of nifti metadata."

df_wide.head()

df_wide

from PIL import Image, ImageOps

def crop_image(row):
    # read in image
    img = Image.open(row['JPG_PATH'].replace('/Documents/Code/', '/PycharmProjects/'))
    
    # crop 
    l = row['bbox_left']
    t = row['bbox_top']
    r = row['bbox_left'] + row['bbox_width']
    b = row['bbox_top'] + row['bbox_height']
    img = img.crop((l, t, r, b))
    
    # reverse negatives
    if row['brightness_flipped']:
        img = ImageOps.invert(img)
    
    # save cropped image
    save_path = cropped_image_path + row['JPG_PATH'].split('/')[-1]
    img.save(save_path)
    
    return save_path

for i, row in df_wide.iterrows():
    print(f"Processing image {i+1}/{len(df)}.", end='\r')
    
    # only process if a valid phf view of good quality
    if not (row['poor_quality'] or row['not_a_phf_view']):
        df_wide.loc[i,'CROPPED_IMAGE_PATH'] = crop_image(row)

df_wide.head()

df_wide.to_csv(cropped_meta_path, index=None)

cropped_meta_path
