==> Xray/Pre-process/3_prep_for_model.ipynb <==
import pandas as pd

img_metadata_path = '../data/cropped_image_metadata.csv'
target_metadata_path = '../data/Fracture_Type.csv'
sbf_view_path = '../data/unnamed_views_sbf_export.csv'

model_img_data_path = '../data/model_images.csv'
model_trg_data_path = '../data/model_targets.csv'

df_img = pd.read_csv(img_metadata_path)
df_trg = pd.read_csv(target_metadata_path)
df_sbf = pd.read_csv(sbf_view_path).rename(columns = {'STD_VIEW': 'view_std_sbf', 'FILENAME_VIEW': 'view'})

df_img

# clean image data
df_img = df_img.dropna(subset='CROPPED_IMAGE_PATH')
df_img = df_img.rename(columns = {'CROPPED_IMAGE_PATH': 'path', 'FILENAME_VIEW': 'view'})
df_img['patient_id'] = df_img['external_id'].apply(lambda x: x.split('_')[0])
df_img = df_img[['patient_id', 'external_id', 'view', 'path']]
df_img

### Record of misclassified studies
AN00017 has both images marked as AP-like unclear why


# standardize view names
df_img['view_std'] = None

# order matters here
patterns = dict(
    G = '.*GRAS.*$', # some mention G and AP. Put first to defer to G
    AP = '^(?!.*\_Y|AXIL|GRAS|LAT|SCAP).*AP.*$',  # some of the names mention both Y and AP
    H =  '^(?!.*AP|AXIL|GRAS|LAT).*HUMER.*$',
    TH = '^(?!.*AP|AXIL|LAT|GRAS|HUMER).*THORAC.*$',
    AX = '^(?!.*AP|HUMER|GRAS|LAT).*AXIL|AXIO.*$',
    Y = '^(?!.*AP|GRAS|AXIL|LAT).*Y.*$',  # Y is last because axillary and grashley both have "Y" in them. Check for false positives
)

for k, v in patterns.items():
    k_cond = df_img.view.str.contains(v, case=False)
    k_cond = k_cond & df_img.view_std.isna()
    df_img.loc[k_cond, 'view_std'] = k
   
    prev_cond = k_cond

# merge on sbf input
dfm = df_img.merge(df_sbf, how='left')

# code special cases from email "Standardized_view_name"
dfm.loc[dfm.view=='LAT_HUMERUS_2', 'view_std'] = 'Y'
dfm.loc[dfm.view=='HUMERUS_LAT_2', 'view_std'] = 'AP'
dfm.loc[dfm.view=='HUMERUS_LAT_3', 'view_std'] = 'AP'

# case 1: std_view is missing and std_view_sbf is not missing:
# action: use sbf
cond_1 = dfm.view_std.isna() & ~dfm.view_std_sbf.isna()
dfm.loc[cond_1, 'view_std'] = dfm.loc[cond_1, 'view_std_sbf']

# case 2: std_view is not missing and std_view_sbf is missing:
# no action taken since STD_VIEW already set appropriately

# case 3 a) both are non-missing and agree
# action: they agree so do nothing

# case 3 b) both are non-missing and they disagree
# action fix on a case-by case basis
cond_3b = (~dfm.view_std.isna() & ~dfm.view_std_sbf.isna()) & (dfm.view_std != dfm.view_std_sbf)
display(dfm.loc[cond_3b, ['view', 'view_std', 'view_std_sbf']])

# case 4: both are missing
# action: after cleaning, only those marked as ambiguous by SBF remain, so going to work with this.
cond_4 = (dfm.view_std.isna() & dfm.view_std_sbf.isna())
dfm.loc[cond_4, ['view', 'view_std', 'view_std_sbf']]

# Add standardized view names to images
dfm = dfm.drop(['view', 'view_std_sbf', 'count'], axis=1)
df_img = dfm.rename(columns = {'view_std': 'view'})

# add high-level view names
group_view_mapping = {
    'AP': 'AP_LIKE',
    'H': 'AP_LIKE',
    'G': 'AP_LIKE',
    'TH': 'AP_LIKE',
    'Y': 'Y',
    'AX': 'AXIL',
}

df_img['view_group'] = df_img.view.map(group_view_mapping)
df_img

df_img.view.value_counts(dropna=False)

df_img.view_group.value_counts(dropna=False)

df_img.query('view_group=="AP_LIKE"').value_counts('view')

# clean target data
df_trg = df_trg.rename(columns = {'AN_': 'patient_id', 'Num_parts': 'num_parts', 'Fracture_type': 'fracture_type'})
df_trg['patient_id'] = df_trg['patient_id'].apply(lambda x: f"AN{str(x).zfill(5)}")
df_trg

# save both sets
df_img.to_csv(model_img_data_path, index=None)
df_trg.to_csv(model_trg_data_path, index=None)

df_img.loc[df_img.external_id.str.contains('lat_humerus_2', case=False)]
